# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface HarAddress {
  /** @nullable */
  streetNumber?: string | null;
  /** @nullable */
  streetName?: string | null;
  /** @nullable */
  streetSuffix?: string | null;
  /** @nullable */
  streetDirection?: string | null;
  /** @nullable */
  unitNumber?: string | null;
  /** @nullable */
  city?: string | null;
  /** @nullable */
  state?: string | null;
  /** @nullable */
  zip?: string | null;
  /** @nullable */
  neighborhood?: string | null;
  /** @nullable */
  area?: string | null;
  /** @nullable */
  country?: string | null;
  /** @nullable */
  addressKey?: string | null;
}

export interface HarPropertyDetails {
  /** @nullable */
  numBedrooms?: number | null;
  /** @nullable */
  numBathrooms?: number | null;
  /** @nullable */
  sqft?: string | null;
  /** @nullable */
  numGarageSpaces?: number | null;
  /** @nullable */
  propertyType?: string | null;
  /** @nullable */
  style?: string | null;
}

export interface HarListingEnrichment {
  address?: HarAddress;
  images?: string[];
  /** @nullable */
  listDate?: string | null;
  /** @nullable */
  soldDate?: string | null;
  /** @nullable */
  listPrice?: number | null;
  /** @nullable */
  soldPrice?: number | null;
  details?: HarPropertyDetails;
}

export type HarViewRecord = HarListingEnrichment & {
  dateTracked: string;
  laAgentkey?: string;
  mlsNumber: string;
  webView: number;
  mobileView: number;
};

export interface HarViewsResponse {
  success: boolean;
  total: number;
  limit: number;
  offset: number;
  data: HarViewRecord[];
}

export type HarReviewRecord = HarListingEnrichment & ({
  mlsNumber: string;
  /** @nullable */
  memberNumber?: string | null;
  /** @nullable */
  agentId?: string | null;
  /** @nullable */
  agentkey?: string | null;
  /** @nullable */
  agentEmail?: string | null;
  /** @nullable */
  clientName?: string | null;
  /** @nullable */
  clientEmail?: string | null;
  /** @nullable */
  transactionType?: string | null;
  /** @nullable */
  score?: number | null;
  /** @nullable */
  competency?: number | null;
  /** @nullable */
  knowledge?: number | null;
  /** @nullable */
  communication?: number | null;
  /** @nullable */
  myExperience?: number | null;
  /** @nullable */
  comments?: string | null;
  /** @nullable */
  agentComment?: string | null;
  /** @nullable */
  closedDate?: string | null;
  /** @nullable */
  publishDate?: string | null;
  /** @nullable */
  createdDate?: string | null;
});

export interface HarReviewsResponse {
  success: boolean;
  total: number;
  limit: number;
  offset: number;
  data: HarReviewRecord[];
}

export interface HarShowingAgent {
  /** @nullable */
  agentkey?: string | null;
  /** @nullable */
  agentName?: string | null;
  /** @nullable */
  agentEmail?: string | null;
  /** @nullable */
  agentPhone?: string | null;
  /** @nullable */
  officeName?: string | null;
}

export interface HarShowingEntry {
  /** @nullable */
  showingId?: string | null;
  /** @nullable */
  showingDate?: string | null;
  /** @nullable */
  showingTime?: string | null;
  /** @nullable */
  duration?: number | null;
  /** @nullable */
  status?: string | null;
  /** @nullable */
  feedback?: string | null;
  buyerAgent?: HarShowingAgent;
}

export type HarShowingRecord = HarListingEnrichment & ({
  mlsNumber: string;
  /** @nullable */
  laAgentkey?: string | null;
  showings?: HarShowingEntry[];
});

export interface HarShowingsResponse {
  success: boolean;
  total: number;
  limit: number;
  offset: number;
  data: HarShowingRecord[];
}

export interface HarSummary {
  totalWebViews: number;
  totalMobileViews: number;
  /** @nullable */
  averageReviewScore?: number | null;
  totalReviews: number;
  totalShowings: number;
  uniqueListingsWithViews: number;
  uniqueListingsWithShowings: number;
}

export type GetHarViewsParams = {
mlsNumber?: string;
dateBegin?: string;
dateEnd?: string;
sortBy?: string;
limit?: number;
offset?: number;
boardAgentId?: string;
};

export type GetHarReviewsParams = {
mlsNumber?: string;
dateBegin?: string;
dateEnd?: string;
sortBy?: string;
limit?: number;
offset?: number;
boardAgentId?: string;
};

export type GetHarShowingsParams = {
mlsNumber?: string;
dateBegin?: string;
dateEnd?: string;
sortBy?: string;
limit?: number;
offset?: number;
};

export type GetHarSummaryParams = {
dateBegin?: string;
dateEnd?: string;
};