# lib/api-zod/src/generated/types/harAddress.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface HarAddress {
  /** @nullable */
  streetNumber?: string | null;
  /** @nullable */
  streetName?: string | null;
  /** @nullable */
  streetSuffix?: string | null;
  /** @nullable */
  streetDirection?: string | null;
  /** @nullable */
  unitNumber?: string | null;
  /** @nullable */
  city?: string | null;
  /** @nullable */
  state?: string | null;
  /** @nullable */
  zip?: string | null;
  /** @nullable */
  neighborhood?: string | null;
  /** @nullable */
  area?: string | null;
  /** @nullable */
  country?: string | null;
  /** @nullable */
  addressKey?: string | null;
}