# lib/api-zod/src/generated/types/harListingEnrichment.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { HarAddress } from './harAddress';
import type { HarPropertyDetails } from './harPropertyDetails';

export interface HarListingEnrichment {
  address?: HarAddress;
  images?: string[];
  /** @nullable */
  listDate?: string | null;
  /** @nullable */
  soldDate?: string | null;
  /** @nullable */
  listPrice?: number | null;
  /** @nullable */
  soldPrice?: number | null;
  details?: HarPropertyDetails;
}