# lib/api-zod/src/generated/types/harPropertyDetails.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface HarPropertyDetails {
  /** @nullable */
  numBedrooms?: number | null;
  /** @nullable */
  numBathrooms?: number | null;
  /** @nullable */
  sqft?: string | null;
  /** @nullable */
  numGarageSpaces?: number | null;
  /** @nullable */
  propertyType?: string | null;
  /** @nullable */
  style?: string | null;
}