# lib/api-zod/src/generated/types/harReviewRecord.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { HarListingEnrichment } from './harListingEnrichment';

export type HarReviewRecord = HarListingEnrichment & ({
  mlsNumber: string;
  /** @nullable */
  memberNumber?: string | null;
  /** @nullable */
  agentId?: string | null;
  /** @nullable */
  agentkey?: string | null;
  /** @nullable */
  agentEmail?: string | null;
  /** @nullable */
  clientName?: string | null;
  /** @nullable */
  clientEmail?: string | null;
  /** @nullable */
  transactionType?: string | null;
  /** @nullable */
  score?: number | null;
  /** @nullable */
  competency?: number | null;
  /** @nullable */
  knowledge?: number | null;
  /** @nullable */
  communication?: number | null;
  /** @nullable */
  myExperience?: number | null;
  /** @nullable */
  comments?: string | null;
  /** @nullable */
  agentComment?: string | null;
  /** @nullable */
  closedDate?: string | null;
  /** @nullable */
  publishDate?: string | null;
  /** @nullable */
  createdDate?: string | null;
});