# lib/api-zod/src/generated/types/harReviewsResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { HarReviewRecord } from './harReviewRecord';

export interface HarReviewsResponse {
  success: boolean;
  total: number;
  limit: number;
  offset: number;
  data: HarReviewRecord[];
}