# lib/api-zod/src/generated/types/harShowingAgent.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface HarShowingAgent {
  /** @nullable */
  agentkey?: string | null;
  /** @nullable */
  agentName?: string | null;
  /** @nullable */
  agentEmail?: string | null;
  /** @nullable */
  agentPhone?: string | null;
  /** @nullable */
  officeName?: string | null;
}