# lib/api-zod/src/generated/types/harShowingEntry.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { HarShowingAgent } from './harShowingAgent';

export interface HarShowingEntry {
  /** @nullable */
  showingId?: string | null;
  /** @nullable */
  showingDate?: string | null;
  /** @nullable */
  showingTime?: string | null;
  /** @nullable */
  duration?: number | null;
  /** @nullable */
  status?: string | null;
  /** @nullable */
  feedback?: string | null;
  buyerAgent?: HarShowingAgent;
}