# lib/api-zod/src/generated/types/harShowingRecord.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { HarListingEnrichment } from './harListingEnrichment';
import type { HarShowingEntry } from './harShowingEntry';

export type HarShowingRecord = HarListingEnrichment & ({
  mlsNumber: string;
  /** @nullable */
  laAgentkey?: string | null;
  showings?: HarShowingEntry[];
});