# lib/api-zod/src/generated/types/harShowingsResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { HarShowingRecord } from './harShowingRecord';

export interface HarShowingsResponse {
  success: boolean;
  total: number;
  limit: number;
  offset: number;
  data: HarShowingRecord[];
}