# lib/api-zod/src/generated/types/harSummary.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface HarSummary {
  totalWebViews: number;
  totalMobileViews: number;
  /** @nullable */
  averageReviewScore?: number | null;
  totalReviews: number;
  totalShowings: number;
  uniqueListingsWithViews: number;
  uniqueListingsWithShowings: number;
}