# lib/api-zod/src/generated/types/harViewRecord.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { HarListingEnrichment } from './harListingEnrichment';

export type HarViewRecord = HarListingEnrichment & {
  dateTracked: string;
  laAgentkey?: string;
  mlsNumber: string;
  webView: number;
  mobileView: number;
};