# lib/api-zod/src/generated/types/harViewsResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { HarViewRecord } from './harViewRecord';

export interface HarViewsResponse {
  success: boolean;
  total: number;
  limit: number;
  offset: number;
  data: HarViewRecord[];
}